chore: remove deprecated helpers
