+++ apps/campfire/src/hooks/useSerializedDirectiveRunner.ts
@@ -3,7 +3,7 @@ import type { RootContent } from 'mdast'
 import type { ContainerDirective } from 'mdast-util-directive'
 import rfdc from 'rfdc'
 import { evalExpression } from '@campfire/utils/evalExpression'
-import { runDirectiveBlock } from '@campfire/utils/directives'
+import { runDirectiveBlock } from '@campfire/utils/directiveHelpers'
 import { useDirectiveHandlers } from '@campfire/hooks/useDirectiveHandlers'
 import { useGameStore } from '@campfire/state/useGameStore'
 import { getLabel, stripLabel } from '@campfire/remark-campfire/helpers'
